refactor: split pagination
