+++ ayon_server/graphql/resolvers/projects.py
@@ -11,9 +11,9 @@ from ayon_server.graphql.resolvers.common import (
     ARGLast,
     FieldInfo,
     argdesc,
-    create_pagination,
     resolve,
 )
+from ayon_server.graphql.resolvers.pagination import create_pagination
 from ayon_server.graphql.types import Info
 from ayon_server.types import validate_name
 from ayon_server.utils import SQLTool
