+++ ayon_server/graphql/resolvers/tasks.py
@@ -17,14 +17,15 @@ from ayon_server.graphql.resolvers.common import (
     FieldInfo,
     argdesc,
     create_folder_access_list,
-    create_pagination,
     get_has_links_conds,
     resolve,
     sortdesc,
 )
+from ayon_server.graphql.resolvers.pagination import (
+    create_pagination,
+    get_attrib_sort_case,
+)
 from ayon_server.graphql.types import Info
-from ayon_server.lib.postgres import Postgres
-from ayon_server.logging import logger
 from ayon_server.sqlfilter import QueryFilter, build_filter
 from ayon_server.types import (
     sanitize_string_list,
@@ -42,23 +43,6 @@ SORT_OPTIONS = {
     "taskType": "tasks.folder_type",
     "assignees": "array_to_string(tasks.assignees, '')",
 }
-
-
-async def get_priority_case() -> str:
-    attr = "priority"
-    res = await Postgres.fetch(
-        "SELECT data->'enum' as enum FROM attributes where name = $1", attr
-    )
-    if not res or not res[0]["enum"]:
-        return f"(pf.attrib || tasks.attrib)->>'{attr}'"
-    case = "CASE"
-    i = 0
-    for i, eval in enumerate(res[0]["enum"]):
-        e = eval["value"]
-        case += f" WHEN (pf.attrib || tasks.attrib)->>'{attr}' = '{e}' THEN {i}"
-    case += f" ELSE {i+1}"
-    case += " END"
-    return case
 
 
 async def get_tasks(
@@ -227,9 +211,6 @@ async def get_tasks(
     elif root.__class__.__name__ == "FolderNode":
         # cannot use isinstance here because of circular imports
         sql_conditions.append(f"tasks.folder_id = '{root.id}'")
-
-    # if name:
-    #     sql_conditions.append(f"tasks.name ILIKE '{name}'")
 
     if names is not None:
         if not names:
@@ -429,12 +410,11 @@ async def get_tasks(
     if sort_by is not None:
         if sort_by in SORT_OPTIONS:
             order_by.insert(0, SORT_OPTIONS[sort_by])
-        if sort_by == "attrib.priority":
-            priority_case = await get_priority_case()
-            order_by.insert(0, priority_case)
         elif sort_by.startswith("attrib."):
-            r = f"(pf.attrib || tasks.attrib)->'{sort_by[7:]}'"  # noqa
-            order_by.insert(0, r)
+            attr_name = sort_by[7:]
+            exp = "(pf.attrib || tasks.attrib)"
+            attr_case = await get_attrib_sort_case(attr_name, exp)
+            order_by.insert(0, attr_case)
         else:
             raise ValueError(f"Invalid sort_by value: {sort_by}")
 
@@ -470,7 +450,7 @@ FROM project_{project_name}.tasks AS tasks
 {ordering}
     """
 
-    logger.debug(f"Task query\n{query}")
+    # logger.debug(f"Task query\n{query}")
 
     return await resolve(
         TasksConnection,
