+++ ayon_server/graphql/resolvers/users.py
@@ -10,9 +10,9 @@ from ayon_server.graphql.resolvers.common import (
     ARGFirst,
     ARGLast,
     argdesc,
-    create_pagination,
     resolve,
 )
+from ayon_server.graphql.resolvers.pagination import create_pagination
 from ayon_server.graphql.types import Info
 from ayon_server.types import validate_name_list, validate_user_name
 from ayon_server.utils import SQLTool
